refactor: 563 - move school menu to dedicated folder
