+++ src/widgets/school-menu/ui/school-menu/school-menu.scss
