+++ src/widgets/school-menu/ui/school-menu/school-menu.test.tsx
@@ -1,6 +1,6 @@
 import { screen } from '@testing-library/react';
 import { vi } from 'vitest';
-import { SchoolMenu } from './ui/school-menu';
+import { SchoolMenu } from '../school-menu/school-menu';
 import { Course } from '@/entities/course';
 import { MOCKED_IMAGE_PATH } from '@/shared/__tests__/constants';
 import { renderWithRouter } from '@/shared/__tests__/utils';
